LayoutTests/imported/w3c/web-platform-tests/css/css-align/abspos/justify-self-default-overflow-vrl-ltr-vrl-expected.txt













PASS .item 1
PASS .item 2
FAIL .item 3 assert_equals:
<div class="container">
  <div class="item" style="top: 20px; bottom: 10px;" data-offset-y="5">
    <div class="inner" style="height: 95px;"></div>
  </div>
</div>
offsetTop expected 5 but got 8
FAIL .item 4 assert_equals:
<div class="container">
  <div class="item" style="top: 20px; bottom: 10px;" data-offset-y="0">
    <div class="inner" style="height: 120px;"></div>
  </div>
</div>
offsetTop expected 0 but got -5
PASS .item 5
PASS .item 6
FAIL .item 7 assert_equals:
<div class="container">
  <div class="item" style="top: -20px; bottom: -10px;" data-offset-y="-20">
    <div class="inner" style="height: 160px;"></div>
  </div>
</div>
offsetTop expected -20 but got -35
PASS .item 8
PASS .item 9
FAIL .item 10 assert_equals:
<div class="container">
  <div class="item rtl" style="top: 20px; bottom: 10px;" data-offset-y="5">
    <div class="inner" style="height: 95px;"></div>
  </div>
</div>
offsetTop expected 5 but got 8
FAIL .item 11 assert_equals:
<div class="container">
  <div class="item rtl" style="top: 20px; bottom: 10px;" data-offset-y="0">
    <div class="inner" style="height: 120px;"></div>
  </div>
</div>
offsetTop expected 0 but got -5
PASS .item 12
PASS .item 13
FAIL .item 14 assert_equals:
<div class="container">
  <div class="item rtl" style="top: -20px; bottom: -10px;" data-offset-y="-20">
    <div class="inner" style="height: 160px;"></div>
  </div>
</div>
offsetTop expected -20 but got -35
